requirements.txt
streamlit
openai
matplotlib
pdfplumber
python-dotenv